extract component for code input
